+++ scripts/macros/2024/classFeatures/rogue/thief/supremeSneak.js
@@ -1,0 +1,5 @@
+export let supremeSneak = {
+    name: 'Supreme Sneak',
+    version: '1.3.61',
+    rules: 'modern'
+};